refactor: Extract paragraph allocator
